Integration test for the default retention period feature
With the default retention period feature, null value for expiration time is a valid value. If the expiration time is not specified (null) at the time of create backup then the backup will be retained for the duration of maximum retention period.

In this commit, an integration test is added for the scenario where the expiration time is null at the time of create. Since the backend changes aren't landed, it is expected to fail with an exception. This would eventually be fixed to succeed.