Copy Backup integration test for the default retention period feature
With the default retention period feature, null value for expiration time is a valid value. If the expiration time is not specified (null) at the time of copy backup then the source backup expiration time will be used as the expiration time for the destination backup as well.

In this commit, an integration test is added for the scenario where the expiration time is null at the time of copy. Since the backend changes aren't landed, it is expected to fail with an exception. This would eventually be fixed to succeed.